configure files so linux will work without redundant  code files
