--- JDownloader 2.0/jd/captcha/methods/keep2share.cc/jacinfo.xml
@@ -1,4 +1,0 @@
-<jDownloader>
-  <method type="extern" os="windows" name="kscaptcha" services="keep2share.cc;fileboom.me;depositfiles.com;" />
-  <command cmd="tools\Windows\offlineCaptchaSolver\keep2share.cc.bat" src="tools\Windows\offlineCaptchaSolver\input.gif" dst="tools\Windows\offlineCaptchaSolver\result.txt" />
-</jDownloader>
