+++ JDownloader 2.0/jd/captcha/methods/keep2share_linux/jacinfo.xml
@@ -1,0 +1,4 @@
+<jDownloader>
+  <method type="extern" os="linux" name="kscaptcha_linux" services="keep2share.cc;fileboom.me;depositfiles.com;" />
+  <command cmd="tools/offlineCaptchaSolver/keep2share.cc.sh" src="tools/offlineCaptchaSolver/input.gif" dst="tools/offlineCaptchaSolver/result.txt" />
+</jDownloader>
