+++ JDownloader 2.0/jd/captcha/methods/keep2share_win/jacinfo.xml
@@ -1,0 +1,4 @@
+<jDownloader>
+  <method type="extern" os="windows" name="kscaptcha_win" services="keep2share.cc;fileboom.me;depositfiles.com;" />
+  <command cmd="tools\offlineCaptchaSolver\keep2share.cc.bat" src="tools\offlineCaptchaSolver\input.gif" dst="tools\offlineCaptchaSolver\result.txt" />
+</jDownloader>
