--- JDownloader 2.0/jd/captcha/methods/przeklej.org/jacinfo.xml
@@ -1,4 +1,0 @@
-<jDownloader>
-  <method type="extern" os="windows" name="przeklej" services="przeklej.org;" />
-  <command cmd="tools\Windows\offlineCaptchaSolver\przeklej.org.bat" src="tools\Windows\offlineCaptchaSolver\input.gif" dst="tools\Windows\offlineCaptchaSolver\result.txt" />
-</jDownloader>
