+++ JDownloader 2.0/jd/captcha/methods/przeklej_linux/jacinfo.xml
@@ -1,0 +1,4 @@
+<jDownloader>
+  <method type="extern" os="linux" name="przeklej_linux" services="przeklej.org;" />
+  <command cmd="tools/offlineCaptchaSolver/przeklej.org.sh" src="tools/offlineCaptchaSolver/input.gif" dst="tools/offlineCaptchaSolver/result.txt" />
+</jDownloader>
