+++ JDownloader 2.0/jd/captcha/methods/przeklej_win/jacinfo.xml
@@ -1,0 +1,4 @@
+<jDownloader>
+  <method type="extern" os="windows" name="przeklej_win" services="przeklej.org;" />
+  <command cmd="tools\offlineCaptchaSolver\przeklej.org.bat" src="tools\offlineCaptchaSolver\input.gif" dst="tools\offlineCaptchaSolver\result.txt" />
+</jDownloader>
