--- JDownloader 2.0/tools/Windows/offlineCaptchaSolver/checkdeps.bat
@@ -1,6 +1,0 @@
-IF EXIST node_modules (
-echo "all good!"
-) ELSE (
-npm i jimp
-npm i tesseract.js
-)
