--- JDownloader 2.0/tools/Windows/offlineCaptchaSolver/keep2share.cc.bat
@@ -1,2 +1,0 @@
-call checkdeps.bat
-node.exe ocr.js keep2share.cc
