--- JDownloader 2.0/tools/Windows/offlineCaptchaSolver/log.txt
@@ -1,1 +1,0 @@
-{"host":"przeklej.org","text":"","confidence":0}
