--- JDownloader 2.0/tools/Windows/offlineCaptchaSolver/przeklej.org.bat
@@ -1,2 +1,0 @@
-call checkdeps.bat
-node.exe ocr.js przeklej.org
