+++ JDownloader 2.0/tools/offlineCaptchaSolver/checkdeps.bat
@@ -1,0 +1,6 @@
+IF EXIST node_modules (
+echo "all good!"
+) ELSE (
+npm i jimp
+npm i tesseract.js
+)
