+++ JDownloader 2.0/tools/offlineCaptchaSolver/checkdeps.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+if [ -d "node_modules" ]; then
+	echo "all good!"
+else
+	npm i jimp
+	npm i tesseract.js
+fi
