+++ JDownloader 2.0/tools/offlineCaptchaSolver/keep2share.cc.bat
@@ -1,0 +1,2 @@
+call checkdeps.bat
+node.exe ocr.js keep2share.cc
