+++ JDownloader 2.0/tools/offlineCaptchaSolver/keep2share.cc.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+./checkdeps.sh
+./node ocr.js keep2share.cc
