+++ JDownloader 2.0/tools/offlineCaptchaSolver/log.txt
@@ -1,0 +1,1 @@
+{"host":"przeklej.org","text":"","confidence":0}
