+++ JDownloader 2.0/tools/offlineCaptchaSolver/przeklej.org.bat
@@ -1,0 +1,2 @@
+call checkdeps.bat
+node.exe ocr.js przeklej.org
