+++ JDownloader 2.0/tools/offlineCaptchaSolver/przeklej.org.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+./checkdeps.sh
+./node ocr.js przeklej.org
