--- linux/JDownloader 2.0/tools/linux/offlineCaptchaSolver/checkdeps.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-if [ -d "node_modules" ]; then
-	echo "all good!"
-else
-	npm i jimp
-	npm i tesseract.js
-fi
