--- linux/JDownloader 2.0/tools/linux/offlineCaptchaSolver/keep2share.cc.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-./checkdeps.sh
-./node ocr.js keep2share.cc
