--- linux/JDownloader 2.0/tools/linux/offlineCaptchaSolver/log.txt
@@ -1,1 +1,0 @@
-{"host":"przeklej.org","text":"","confidence":0}
