--- linux/JDownloader 2.0/tools/linux/offlineCaptchaSolver/przeklej.org.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-./checkdeps.sh
-./node ocr.js przeklej.org
